% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{startFactorViz}
\alias{startFactorViz}
\title{startFactorViz
This function intialises factorviz and start}
\usage{
startFactorViz(path = NULL)
}
\arguments{
\item{path}{takes in the directory path to decomp output}
}
\description{
startFactorViz
This function intialises factorviz and start
}
