% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{startFactorViz}
\alias{startFactorViz}
\title{startFactorViz
This function intialises factorviz and start}
\usage{
startFactorViz(decomp_output = NULL, medecom_set = NULL,
  ann_C = NULL, ann_S = NULL, ref_meth = NULL)
}
\arguments{
\item{decomp_output}{takes in the directory path to decomp output}

\item{medecom_set}{takes in the path to medecom set file or the MeDeCom Set object}

\item{ann_C}{takes in the path to CpG Annotation file or the data.frame of CpG Annotations}

\item{ann_S}{takes in the path to Sample Annotation file or the data.frame of Sample Annotations}

\item{ref_meth}{takes in the path to Reference Methylome file or the data.frame of Reference Methylome}
}
\description{
startFactorViz
This function intialises factorviz and start
}
\details{
If decomp_output provided all other parameters force set to NULL
}
